Fix and improve build scripts

+++ compile.sh
@@ -80,14 +80,16 @@ fi
 
 # Remove files from previous build
 
-if [[ -d "$cmake_dir" ]] || [[ -d "$app_dir" ]] || [[ -d "$m64p" ]] ; then
+if [[ -d "$cmake_dir" ]] || [[ -d "$app_dir" ]] || [[ -d "$m64p_dir" ]] ; then
+
     echo "One or more working directories (CMake, app, m64p) already exist. Delete folders to proceed?"
     select yn in "Continue" "Exit"; do
         case $yn in
-            Continue ) rm -rf "$cmake_dir" "$app_dir" "$m64p"; break;;
+            Continue ) rm -rf "$cmake_dir" "$app_dir" "$m64p_dir"; break;;
             Exit ) exit;;
         esac
     done
+
 fi
 
 
@@ -102,16 +104,23 @@ mkdir -p "$build_dir" "$cmake_dir" "$emg_dir"
 
 if [[ ! $(uname -s) = Darwin* && ! -f "$emg_dir"/EMG$exe ]] || [[ $(uname -s) = Darwin* && ! -d "$emg_dir"/EMG.app ]]; then
 
-    [[ ! -f "$build_dir"/electron-$electron-$platform-x64.zip ]] && wget https://github.com/electron/electron/releases/download/$electron/electron-$electron-$platform-x64.zip -P "$build_dir"/
+    if [[ ! -f "$build_dir"/electron-$electron-$platform-x64.zip ]]; then
+
+        wget https://github.com/electron/electron/releases/download/$electron/electron-$electron-$platform-x64.zip -P "$build_dir"/
+
+    fi
 
     if [[ $(uname -s) = MINGW64* ]] || [[ $(uname -s) = Linux* ]]; then
+
         unzip -o "$build_dir"/electron-$electron-$platform-x64.zip -d "$build_dir"/EMG -x LICENSE LICENSES.chromium.html version resources/default_app.asar 'locales/*.pak'
         unzip -o "$build_dir"/electron-$electron-$platform-x64.zip locales/en-US.pak -d "$build_dir"/EMG
 
         mv "$emg_dir"/electron$exe "$emg_dir"/EMG$exe
+
     fi
 
     if [[ $(uname -s) = Darwin* ]]; then
+
         unzip -o "$build_dir"/electron-$electron-$platform-x64.zip -d "$build_dir"/EMG -x LICENSE LICENSES.chromium.html version Electron.app/Contents/Resources/default_app.asar Electron.app/Contents/Resources/electron.icns
 
         cp "$info_dir"/EMG.plist "$electron_dir"/Info.plist
@@ -124,8 +133,11 @@ if [[ ! $(uname -s) = Darwin* && ! -f "$emg_dir"/EMG$exe ]] || [[ $(uname -s) = 
     fi
 
     if [[ $(uname -s) = MINGW64* ]]; then
+
         cmd //c "$tools_dir"/rcedit-x64 "$emg_dir"/EMG.exe --set-icon "$ico_dir"/emg.ico --set-version-string LegalCopyright "(C) 2024 EvilGames.eu" --set-version-string OriginalFilename "electron.exe" --set-version-string FileDescription "EMG" --set-version-string ProductName "EMG" --set-version-string CompanyName "EvilGames.eu"
+
     fi
+
 fi
 
 
@@ -141,7 +153,9 @@ cp -R "$toplvl_dir"/resources/* "$rsc_dir"
 # Fix getRevision.sh file permissions
 
 if [[ $(uname -s) = Linux* ]] || [[ $(uname -s) = Darwin* ]]; then
+
     chmod u+x "$toplvl_dir"/source/mupen64plus-video-GLideN64/src/getRevision.sh
+
 fi
 
 
@@ -153,21 +167,28 @@ cmake --build "$cmake_dir" --parallel "$threads"
 cmake --install "$cmake_dir"
 
 if [[ $(uname -s) = MINGW64* ]]; then
+
     cmake --build "$cmake_dir" --target=bundle_dependencies
+
 fi
 
 
 
 # Rename files
 
-mv "$plugin_dir"/$gca "$plugin_dir"/mupen64plus-input-gca$ext
+if [[ -f "$plugin_dir"/$gca ]]; then
 
+    mv "$plugin_dir"/$gca "$plugin_dir"/mupen64plus-input-gca$ext
+
+fi
 
 
 # Add icon to mupen64plus.exe
 
 if [[ $(uname -s) = MINGW64* ]]; then
+
     cmd //c "$tools_dir"/rcedit-x64 "$m64p_dir"/mupen64plus.exe --set-icon "$ico_dir"/mupen64plus.ico
+
 fi
 
 
@@ -175,8 +196,10 @@ fi
 # Fix executable file permissions
 
 if [[ $(uname -s) = Linux* ]] || [[ $(uname -s) = Darwin* ]]; then
+
     chmod u+x "$m64p_dir"/mupen64plus
     chmod u+x "$m64p_dir"/sdl2-jstest
+
 fi
 
 
@@ -184,6 +207,8 @@ fi
 # Strip binaries
 
 if [[ $(uname -s) = MINGW64* ]] || [[ $(uname -s) = Linux* ]]; then
+
     for f in "$m64p_dir"/*$ext; do strip --strip-unneeded $f; done
     for f in "$plugin_dir"/*$ext; do strip --strip-unneeded $f; done
+
 fi
