+++ rt64.sh
@@ -32,7 +32,9 @@ rt64_release_dir="$rt64_build_dir/Release"
 ext=".so"
 
 if [[ $(uname -s) = MINGW64* ]]; then
+
     ext=".dll"
+
 fi
 
 
@@ -40,7 +42,9 @@ fi
 # Clone rt64 recursively
 
 if [[ ! -d "$rt64_dir" ]]; then
+
     git clone https://github.com/GhostlyDark/rt64 "$rt64_dir" --recursive -b plugin --depth 1
+
 fi
 
 
@@ -48,6 +52,7 @@ fi
 # Remove previous build files
 
 if [[ -d "$rt64_build_dir" ]]; then
+
     echo "Build directory for rt64 already exist. Delete folder to proceed?"
     select yn in "Continue" "Exit"; do
         case $yn in
@@ -55,6 +60,7 @@ if [[ -d "$rt64_build_dir" ]]; then
             Exit ) exit;;
         esac
     done
+
 fi
 
 
@@ -64,13 +70,17 @@ fi
 mkdir -p "$rt64_build_dir"
 
 if [[ $(uname -s) = MINGW64* ]]; then
+
     cmake -S "$rt64_dir" -B "$rt64_build_dir" -G "Visual Studio 17 2022"
     cmake --build "$rt64_build_dir" --config Release --parallel "$threads"
+
 fi
 
 if [[ $(uname -s) = Linux* ]]; then
+
     cmake -S "$rt64_dir" -B "$rt64_build_dir" -DCMAKE_BUILD_TYPE=Release
     cmake --build "$rt64_build_dir" --parallel "$threads"
+
 fi
 
 
@@ -78,16 +88,24 @@ fi
 # Copy files
 
 if [[ $(uname -s) = MINGW64* ]]; then
+
     cp "$rt64_build_dir"/dxcompiler.dll "$rt64_build_dir"/dxil.dll "$m64p_dir"
     cp "$rt64_build_dir"/dxcompiler.dll "$rt64_build_dir"/dxil.dll "$rt64_release_dir"/rt64$ext "$plugin_dir"
+
 fi
 
 if [[ $(uname -s) = Linux* ]]; then
+
     cp "$rt64_build_dir"/rt64$ext "$plugin_dir"
+
 fi
 
 
 
 # Rename rt64 to mupen64plus-video-rt64
 
-mv "$plugin_dir"/rt64$ext "$plugin_dir"/mupen64plus-video-rt64$ext
+if [[ -f "$plugin_dir"/rt64$ext ]]; then
+
+    mv "$plugin_dir"/rt64$ext "$plugin_dir"/mupen64plus-video-rt64$ext
+
+fi
